rキーで撮影するようにしました。被験者に促すメッセージを表示するようにしました。
これに伴い、撮影枚数は8枚と固定されます。

+++ main.cpp
@@ -1,31 +1,35 @@
 #include"CKFaceKit.h"
 #include <unistd.h>
-void cap_image(Mat &in,string path, int width, int height,int num_images) {
+void cap_image(Mat &in,string path, int width, int height,int num_images,VideoCapture cap) {
 	int count=0;
-    VideoCapture cap(0);
-    while(!cap.isOpened()){
-        VideoCapture cap(0);
-        printf("カメラの起動中...\n");
-    } 
-    
+    int msgf=0;
+    int fontType = FONT_HERSHEY_SIMPLEX;
 	VideoWriter writer(path, 0, 0.0, Size(width, height));
-    
+    vector<string> msgList={"【正面】","【少し左】","【少し右】","【右上】","【左上】","【左下】","【右下】","【下】"};
+    string msg="を向いて'r'を押してください。";
+
 	while(count<=num_images){	
 		cap >> in;
-		namedWindow("FaceMaker_Capturing",0);
-		setWindowProperty("FaceMaker_Capturing",CV_WND_PROP_FULLSCREEN,CV_WINDOW_FULLSCREEN);
+		namedWindow("FaceMaker_Capturing", CV_WINDOW_AUTOSIZE);
+        if(!msgf){
+            string buf=msgList[count]+msg;
+            printf("%s\n",buf.c_str());
+            msgf=1;
+        }
+		//setWindowProperty("FaceMaker_Capturing",CV_WND_PROP_FULLSCREEN,CV_WINDOW_FULLSCREEN);
 		imshow("FaceMaker_Capturing",in);
-		writer << in;
-		count++;
-		waitKey(30);
-        usleep(500000);
-		}
+        if('r'==waitKey(30)){
+            writer << in;
+            count++;
+            msgf=0;
+        }
+	}
 	destroyWindow("FaceMaker_Capturing");
 	cap.release();
 }
 
 int main(){
-    int PICT=0;
+    int PICT=7;
     char Name[NAME_SIZE];
     char mkcmd[SYSCMD];//クライアント側の./face/にNameのディレクトリ作成
     char sendcmd[SYSCMD];//画像をサーバに送る
@@ -35,14 +39,20 @@ int main(){
     int ws=2;//遅延時間
     string output_path;
     Mat frame,input_image;
-
+    VideoCapture cap(0);
+    if(!cap.isOpened())//カメラデバイスが正常にオープンしたか確認．
+    {
+        //読み込みに失敗したときの処理
+        printf("カメラの起動中\n");
+        return -1;
+    }
     printf("FaceMakerへようこそ\n");
     printf("サーバ側の「face/」ディレクトリを初期化しています...\n");
     sprintf(svrm,"ssh Neptune 'rm -rf FaceMakerServer/face/inpic/*;rm -rf FaceMakerServer/face/outpic/*'");
 	system(svrm);
 
-    printf("撮影枚数:");
-    scanf("%d",&PICT);
+    // printf("撮影枚数:");
+    // scanf("%d",&PICT);
 
     printf("名前:");
     scanf("%s",Name);
@@ -53,7 +63,7 @@ int main(){
     strcat(out_path,"/in_%03d.jpg");
     output_path=string(out_path);
     system(mkcmd);
-    cap_image(frame,output_path, WIDTH,HEIGHT,PICT);//撮影
+    cap_image(frame,output_path, WIDTH,HEIGHT,PICT,cap);//撮影
     sprintf(sendcmd,"scp -r pic/%s Neptune:FaceMakerServer/face/inpic/",Name);
     printf("＊＊Convert＊＊\n");
     printf("サーバに送信しています...\n");
